rename: 범용 fsd 구조에 맞게 구조 변경
