+++ src/features/onboarding/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-onboarding-form.hook";
